dev_requirements.txt
pytest
pytest-cov
pytest-django
pytest-pythonpath